Moved leaderboard to subcommands
I came up with the idea to have `subcommands` when trying to figure out what to do with `target`.

This way, the main command methods can stay clean to their core purpose, but we can have more functionality in each one.
